% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{apercu}
\alias{apercu}
\title{apercu}
\usage{
apercu(o, l = lapply(seq_along(dim(o)), function(x) 1:5))
}
\arguments{
\item{o}{the object for which I need to have the apercu}

\item{l}{the limits}

\item{c}{the classes}
}
\value{
The apercu
}
\description{
apercu
}
\keyword{internal}
