% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{aperWrapper}
\alias{aperWrapper}
\title{aperWrapper}
\usage{
aperWrapper(ob, li = lapply(seq_along(dim(ob)), function(x) 1:5))
}
\arguments{
\item{ob}{the object sent to apercu}

\item{li}{the limits}
}
\value{
the apercu object, of class ap
}
\description{
aperWrapper
}
