% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{classDeterminer}
\alias{classDeterminer}
\title{classDeterminer}
\usage{
classDeterminer(obj)
}
\arguments{
\item{obj}{the object for which the classes have to be determined}
}
\value{
the classes of the object and its elements if list
}
\description{
classDeterminer
}
