% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{fittingLimits}
\alias{fittingLimits}
\title{fittingLimits}
\usage{
fittingLimits(ob, lim)
}
\arguments{
\item{ob}{object with dimensions}

\item{lim}{limits that we have to check}
}
\value{
the corrected limits
}
\description{
fittingLimits
}
\keyword{internal}
