% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{limitsLister}
\alias{limitsLister}
\title{limitsLister}
\usage{
limitsLister(object)
}
\arguments{
\item{object}{The object for which the dimensions should be checked}
}
\value{
A list of the dimensions if none is given
}
\description{
limitsLister
}
\keyword{internal}
