% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap.R
\name{print.ap}
\alias{print.ap}
\title{print.ap}
\usage{
\method{print}{ap}(x, all = FALSE)
}
\arguments{
\item{x}{the ap class object that has to be printed}

\item{all}{should all the elements be printed ?}
}
\value{
prints the object, only the apercu or with dimensions and classes
}
\description{
print.ap
}
